+++ src/Pages/App/App.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Container } from 'react-bootstrap'
 import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom'
-import DiaryHome from './DiaryHome'
-import LoginPage from './LoginPage'
-import Header from '../Components/Header'
+import DiaryHome from '../DiaryHome/DiaryHome'
+import Login from '../Login/Login'
+import Header from '../../Components/Header'
 import { useSelector } from 'react-redux'
-import NotFoundPage from './NotFount'
+import NotFoundPage from '../NotFound/NotFount'
 
 function App () {
   const login = useSelector(state => state.login)
@@ -14,7 +14,7 @@ function App () {
         <Header/>
         <Container fluid className={'vh-100 main no-scroll'}>
           <Switch>
-            <Route exact path='/login'><LoginPage/></Route>
+            <Route exact path='/login'><Login/></Route>
             <Route exact path='/'>
               { login.isLoggedIn ? <DiaryHome/> : <Redirect to='/login'/> }
             </Route>
