+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import './Utils/index.css'
-import App from './Pages/App'
+import './utils/index.css'
+import App from './Pages/App/App'
 import reportWebVitals from './reportWebVitals'
 import 'bootstrap/dist/css/bootstrap.min.css'
-import { store, rrfProps } from './Utils/store'
+import { store, rrfProps } from './utils/store'
 import { Provider } from 'react-redux'
 import { ReactReduxFirebaseProvider } from 'react-redux-firebase'
