+++ src/utils/firebase.js
@@ -1,4 +1,7 @@
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+import firebase from 'firebase/app'
+import 'firebase/auth'
+import 'firebase/firestore'
+
 const firebaseConfig = {
   apiKey: 'AIzaSyBiwWSCcLhdaWU72mbQVgP0GOoyfVMkpIs',
   authDomain: 'dear-diary-salitha.firebaseapp.com',
@@ -9,4 +12,12 @@ const firebaseConfig = {
   measurementId: 'G-E5585KB4EV'
 }
 
-export default firebaseConfig
+firebase.initializeApp(firebaseConfig)
+firebase.firestore()
+
+const rrfConfig = {
+  userProfile: 'users',
+  useFirestoreForProfile: true
+}
+
+export { rrfConfig, firebase }
