+++ src/utils/store.js
@@ -1,16 +1,16 @@
 import { applyMiddleware, createStore } from 'redux'
 import createSagaMiddleware from 'redux-saga'
-import rootReducer from './reducers'
-import { watchSaga } from './sagas/rootSaga'
-import { composeWithDevTools } from 'redux-devtools-extension'
 import { createFirestoreInstance } from 'redux-firestore'
+import { composeWithDevTools } from 'redux-devtools-extension'
+import { watchSaga } from './rootSaga'
+import rootReducer from './rootReducer'
 import { firebase, rrfConfig } from './firebase'
 
 const sagaMiddleware = createSagaMiddleware()
 
 const middlewares = [sagaMiddleware]
 
-export const store = createStore(
+const store = createStore(
   rootReducer,
   composeWithDevTools(applyMiddleware(...middlewares))
 )
@@ -24,4 +24,6 @@ export const rrfProps = {
   createFirestoreInstance // <- needed if using firestore
 }
 
-export const firestore = firebase.firestore()
+const firestore = firebase.firestore()
+
+export { store, firestore }
